requirements.txt
# Vehicle Detection Project - Dependencies
# ==========================================
# Install with: pip install -r requirements.txt
#
# NOTE ON PyTorch:
# - For GPU (CUDA) support, install PyTorch separately first:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# - For CPU-only:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
# - Then install the remaining dependencies below.
# - Visit https://pytorch.org/get-started/locally/ for your specific setup.

# Core ML Frameworks (install PyTorch separately as noted above)
# torch
# torchvision
# torchaudio

# YOLOv11 (Ultralytics)
ultralytics>=8.3.0

# Utilities
tqdm>=4.60.0
Pillow>=9.0.0
matplotlib>=3.5.0
pyyaml>=6.0
opencv-python>=4.5.0
flask>=2.2.0
werkzeug>=2.2.0
